requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
mysql-connector-python==8.4.0
pydantic==2.9.2
spacy==3.7.2
sentence-transformers==3.1.1
python-docx==1.1.0
PyPDF2==3.0.1
langdetect==1.0.9
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.31.1
pytest==8.3.3
scikit-learn==1.4.1.post1